pyproject.toml
[tool.poetry]
name = "mutiSateChatInOrbit"
version = "0.1.0"
description = "Remote sensing large model in orbit inference for multi-satellite collaboration"
authors = ["LucazznS <LucazznS@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
# 添加其他依赖项
# 移出锁 rm poetry.lock
# 锁定依赖版本 poetry lock
# poetry add matplotlib==3.9.2
absl-py = "2.1.0"
accelerate = "0.21.0"
aiofiles = "24.1.0"
aiohappyeyeballs = "2.4.4"
aiohttp = "3.11.10"
aiosignal = "1.3.1"
altair = "5.4.1"
anyio = "4.7.0"
astropy = "6.1.7"
astropy-iers-data = "0.2024.12.16.0.35.48"
async-timeout = "5.0.1"
attrs = "24.2.0"
bidict = "0.23.1"
bitsandbytes = "0.41.0"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
click = "8.1.7"
cmake = "3.31.1"
contourpy = "1.3.1"
cycler = "0.12.1"
deepspeed = "0.9.5"
docker-pycreds = "0.4.0"
einops = "0.6.1"
einops-exts = "0.0.4"
exceptiongroup = "1.2.2"
fastapi = "0.115.0"
fastlogging = "1.2.0"
ffmpy = "0.4.0"
filelock = "3.16.1"
flash-attn = "2.6.3"
fonttools = "4.54.1"
frozenlist = "1.4.1"
fsspec = "2024.9.0"
# geochat = { git = "https://github.com/mbzuai-oryx/GeoChat.git", rev = "045b187803f3f56435d6516bbb1ebef787575326" }
gitdb = "4.0.11"
GitPython = "3.1.43"
gradio = "3.35.2"
gradio_client = "0.2.9"
grpcio = "1.67.0"
h11 = "0.14.0"
hf_transfer = "0.1.8"
hjson = "3.1.0"
httpcore = "0.17.3"
httpx = "0.24.0"
huggingface = "0.0.1"
huggingface-hub = "0.25.1"
idna = "3.10"
inquirerpy = "0.3.4"
itur = "0.4.0"
Jinja2 = "3.1.4"
joblib = "1.4.2"
jplephem = "2.22"
jsonschema = "4.23.0"
jsonschema-specifications = "2023.12.1"
kiwisolver = "1.4.7"
latex2mathml = "3.77.0"
linkify-it-py = "2.0.3"
lit = "18.1.8"
Mako = "1.3.6"
Markdown = "3.7"
markdown-it-py = "2.2.0"
markdown2 = "2.5.0"
MarkupSafe = "2.1.5"
# matplotlib = "3.9.2"
mdit-py-plugins = "0.3.3"
mdurl = "0.1.2"
mpmath = "1.3.0"
multidict = "6.1.0"
narwhals = "1.8.3"
networkx = "3.3"
ninja = "1.11.1.1"
numpy = "1.26.4"
nvidia-cublas-cu11 = "11.11.3.6"
nvidia-cuda-cupti-cu11 = "11.8.87"
nvidia-cuda-nvrtc-cu11 = "11.8.89"
nvidia-cuda-runtime-cu11 = "11.8.89"
nvidia-cudnn-cu11 = "9.1.0.70"
nvidia-cufft-cu11 = "10.9.0.58"
nvidia-curand-cu11 = "10.3.0.86"
nvidia-cusolver-cu11 = "11.4.1.48"
nvidia-cusparse-cu11 = "11.7.5.86"
nvidia-nccl-cu11 = "2.21.5"
nvidia-nvtx-cu11 = "11.8.86"
opencv-python = "4.10.0.84"
orjson = "3.10.7"
packaging = "24.2"
pandas = "2.2.3"
peft = "0.4.0"
pfzy = "0.3.4"
pillow = "10.4.0"
platformdirs = "4.3.6"
prompt_toolkit = "3.0.48"
propcache = "0.2.1"
protobuf = "5.29.1"
psutil = "6.1.0"
py-cpuinfo = "9.0.0"
PyAstronomy = "0.22.0"
pycuda = "2024.1.2"
pydantic = "1.10.19"
pydub = "0.25.1"
pyerfa = "2.0.1.5"
Pygments = "2.18.0"
pyparsing = "3.2.0"
pyproj = "3.7.0"
python-dateutil = "2.9.0.post0"
python-multipart = "0.0.19"
pytools = "2024.1.21"
pytz = "2024.2"
PyYAML = "6.0.2"
quantities = "0.16.1"
referencing = "0.35.1"
regex = "2024.11.6"
requests = "2.32.3"
rpds-py = "0.22.3"
safetensors = "0.4.5"
scikit-learn = "1.2.2"
scipy = "1.14.1"
semantic-version = "2.10.0"
sentencepiece = "0.1.99"
sentry-sdk = "2.19.2"
setproctitle = "1.3.4"
sgp4 = "2.23"
shapely = "2.0.6"
shortuuid = "1.0.13"
six = "1.17.0"
skyfield = "1.49"
smmap = "5.0.1"
sniffio = "1.3.1"
starlette = "0.38.6"
svgwrite = "1.4.3"
sympy = "1.13.1"
tensorboard = "2.18.0"
tensorboard-data-server = "0.7.2"
threadpoolctl = "3.5.0"
timm = "0.6.13"
tokenizers = "0.13.3"
torch = "2.5.1+cu118"
torch-tb-profiler = "0.4.3"
torchaudio = "2.5.1+cu118"
torchvision = "0.20.1+cu118"
tqdm = "4.67.1"
transformers = "4.31.0"
triton = "3.1.0"
typing_extensions = "4.12.2"
tzdata = "2024.2"
uc-micro-py = "1.0.3"
urllib3 = "2.2.3"
uvicorn = "0.32.1"
wandb = "0.18.7"
wavedrom = "2.0.3.post3"
wcwidth = "0.2.13"
websockets = "14.1"
Werkzeug = "3.1.3"
yarl = "1.18.3"
matplotlib = "3.9.2"
